+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInListener.cs
@@ -53,9 +53,18 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation
                 return;
             }
 
-            if (this.options.RequestFilter?.Invoke(context) == false)
+            try
             {
-                AspNetInstrumentationEventSource.Log.RequestIsFilteredOut(activity.OperationName);
+                if (this.options.Filter?.Invoke(context) == false)
+                {
+                    AspNetInstrumentationEventSource.Log.RequestIsFilteredOut(activity.OperationName);
+                    activity.IsAllDataRequested = false;
+                    return;
+                }
+            }
+            catch (Exception ex)
+            {
+                AspNetInstrumentationEventSource.Log.RequestFilterException(ex);
                 activity.IsAllDataRequested = false;
                 return;
             }
@@ -63,7 +72,7 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation
             var request = context.Request;
             var requestValues = request.Unvalidated;
 
-            if (!(this.options.TextFormat is TraceContextFormat))
+            if (!(this.options.TextFormat is TextMapPropagator))
             {
                 var ctx = this.options.TextFormat.Extract(default, request, HttpRequestHeaderValuesGetter);
 
@@ -97,9 +106,7 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation
             var path = requestValues.Path;
             activity.DisplayName = path;
 
-            activity.SetKind(ActivityKind.Server);
-
-            this.activitySource.Start(activity);
+            this.activitySource.Start(activity, ActivityKind.Server);
 
             if (activity.IsAllDataRequested)
             {
@@ -125,7 +132,7 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation
             Activity activityToEnrich = activity;
             Activity createdActivity = null;
 
-            if (!(this.options.TextFormat is TraceContextFormat))
+            if (!(this.options.TextFormat is TextMapPropagator))
             {
                 // If using custom context propagator, then the activity here
                 // could be either the one from Asp.Net, or the one
@@ -190,7 +197,7 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation
                 }
             }
 
-            if (!(this.options.TextFormat is TraceContextFormat))
+            if (!(this.options.TextFormat is TextMapPropagator))
             {
                 if (activity.OperationName.Equals(ActivityNameByHttpInListener))
                 {